requirements.txt
tensorflow
Pillow
requests